// src/store/userStore/action.enum.ts
enum actionTypes {
  AUTH_PENDING = "LOGIN_PENDING",
  AUTH_SUCCESS = "AUTH_SUCCESS",
  LOGOUT = "LOGOUT",
  LOGIN_SUCCESS = "LOGIN__SUCCESS",
  REGISTER_SUCCESS = " REGISTER__SUCCESS",
  FORGOT_PASSWORD_SUCCESS = "FORGOT_PASSWORD__SUCCESS",
  AUTH_ERROR = "AUTH_ERROR",
  USER_LOADED = "USER_LOADED",
  SUCCESS = "SUCCESS",
  CLEAR_SUCCESS = "CLEAR_SUCCESS",
  RESET_PASSWORD_SUCCESS = "RESET_PASSWORD_SUCCESS",
  CHANGE_PASSWORD_SUCCESS = "CHANGE_PASSWORD_SUCCESS",
}

export default actionTypes;